backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-jose
passlib[bcrypt]
psycopg2-binary
alembic